refactor: move ensureCacheReady to tools/shared.ts, fix race timer
1. Move ensureCacheReady from cache.ts to tools/shared.ts — cache
   module should not depend on ToolResult/errorResult (presentation
   layer). CacheReadyCheckable interface stays with the helper.

2. Use deadline-based remaining time for Promise.race timer instead
   of full timeoutMs, preventing any theoretical nanosecond drift.

3. Add JSDoc comment about discarded-build behavior in concurrent
   initialize() caller path.
